fix(postgres-sink): bound checkpoint truncation too; honor disabled timeout
Adversarial review of the first commit surfaced two gaps, both fixed by
moving the bounded-execute guard into streamling-core where every
postgres statement path can share it:

1. The checkpoint-truncation DELETE runs inline in the sink's batch
   loop (enabled for every aggregation sink) with no retry wrapper —
   the identical silent hang this PR exists to kill. It now routes
   through execute_bounded.

2. statement_timeout_secs = 0 documented 'no timeout' but the first
   commit mapped it to a hard 300s client bound, which turns any
   legitimately-slower statement into an infinite kill/retry livelock
   with no escape hatch. The client bound is now Option-typed: 0
   disables it too, preserving the contract; every real configuration
   (default 60s) keeps full protection.

Also hardens the unit test based on mutation findings: it now asserts
the bound actually elapsed with the timeout message (an acquire
failure no longer satisfies it) and runs a third attempt against a
2-connection pool, which fails with PoolTimedOut if a regression
replaces detach() with a plain drop (the leaked permit hides in a
background return-to-pool ping).
