+++ crates/streamling-config/src/app_config.rs
@@ -555,15 +555,18 @@ impl PostgresSinkConfig {
     /// the connection is dead (e.g. a middlebox silently dropped the flow), so
     /// sinks additionally bound the await client-side. Derived as 2x the
     /// server-side timeout so the server-side timeout fires first whenever the
-    /// server is reachable; when the server-side timeout is disabled (`0`),
-    /// falls back to 5 minutes.
-    pub fn client_statement_timeout(&self) -> std::time::Duration {
-        const DISABLED_FALLBACK_SECS: u64 = 300;
-        let secs = match self.statement_timeout_secs {
-            0 => DISABLED_FALLBACK_SECS,
-            s => s.saturating_mul(2),
-        };
-        std::time::Duration::from_secs(secs)
+    /// server is reachable (the 2x headroom also covers wire upload time,
+    /// which the server-side timeout does not measure).
+    ///
+    /// `statement_timeout_secs = 0` means "no timeout" and is honored here
+    /// too (`None`): a statement legitimately slower than any fixed bound
+    /// would otherwise be killed and retried forever. Users who disable the
+    /// server timeout opt out of the client bound as well.
+    pub fn client_statement_timeout(&self) -> Option<std::time::Duration> {
+        match self.statement_timeout_secs {
+            0 => None,
+            s => Some(std::time::Duration::from_secs(s.saturating_mul(2))),
+        }
     }
 }
 
@@ -1199,20 +1202,17 @@ password: ""
         };
         assert_eq!(
             config.client_statement_timeout(),
-            std::time::Duration::from_secs(120)
+            Some(std::time::Duration::from_secs(120))
         );
     }
 
     #[test]
-    fn test_client_statement_timeout_falls_back_when_server_timeout_disabled() {
+    fn test_client_statement_timeout_disabled_when_server_timeout_disabled() {
         let config = PostgresSinkConfig {
             statement_timeout_secs: 0,
             ..Default::default()
         };
-        assert_eq!(
-            config.client_statement_timeout(),
-            std::time::Duration::from_secs(300)
-        );
+        assert_eq!(config.client_statement_timeout(), None);
     }
 
     #[test]
@@ -1223,7 +1223,7 @@ password: ""
         };
         assert_eq!(
             config.client_statement_timeout(),
-            std::time::Duration::from_secs(u64::MAX)
+            Some(std::time::Duration::from_secs(u64::MAX))
         );
     }
 }
