+++ crates/streamling-core/src/utils/pg.rs
@@ -1,5 +1,5 @@
 use crate::data::COLUMN_NAME_OP;
-use crate::error::{Result, StreamlingError};
+use crate::error::{Result, ResultExt, StreamlingError};
 use crate::types::{i256::I256Type, u256::U256Type};
 use crate::utils::parse_primary_key_columns;
 use arrow_schema::{DataType, Field, Schema, SchemaRef};
@@ -99,6 +99,60 @@ impl PostgresConnection {
     /// Get a reference to the pool
     pub fn pool(&self) -> &PgPool {
         &self.pool
+    }
+}
+
+/// Execute a query on a pooled connection with an optional client-side time
+/// bound, returning the number of rows affected.
+///
+/// The server-side `statement_timeout` cannot fire when the connection is dead
+/// (e.g. a NAT/firewall silently dropped the flow after the statement was
+/// sent): the client just awaits a response that will never arrive, and sqlx
+/// exposes no TCP keepalive to detect it. Without this bound such an await
+/// hangs the caller forever with no logs. `None` (from
+/// `statement_timeout_secs = 0`) disables the bound — matching the server-side
+/// "no timeout" contract.
+///
+/// On timeout the connection is detached from the pool and dropped rather
+/// than returned: the next acquire would otherwise receive the same dead
+/// socket, and sqlx's return-to-pool cleanup can itself block on it. Dropping
+/// a detached connection closes the socket without any await, and the pool
+/// opens a fresh replacement on the next acquire.
+///
+/// Delivery semantics: like every retried error on these paths (e.g. a
+/// connection drop after the server committed but before the response
+/// arrived), a timeout can lead to re-executing a statement that already
+/// committed — at-least-once. Upsert/delete statements absorb this; callers
+/// issuing plain INSERTs (no conflict target) can observe duplicates, which is
+/// their pre-existing retry contract, not a property of this bound.
+pub async fn execute_bounded(
+    pool: &PgPool,
+    query: sqlx::query::Query<'_, sqlx::Postgres, sqlx::postgres::PgArguments>,
+    timeout: Option<Duration>,
+) -> Result<u64> {
+    let Some(timeout) = timeout else {
+        let result = query
+            .execute(pool)
+            .await
+            .streamling_context("failed to execute statement")?;
+        return Ok(result.rows_affected());
+    };
+
+    let mut conn = pool.acquire().await.map_err(|e| {
+        StreamlingError::retriable_with_cause("failed to acquire PostgreSQL connection", e)
+    })?;
+
+    match tokio::time::timeout(timeout, query.execute(&mut *conn)).await {
+        Ok(result) => {
+            let result = result.streamling_context("failed to execute statement")?;
+            Ok(result.rows_affected())
+        }
+        Err(_elapsed) => {
+            drop(conn.detach());
+            Err(StreamlingError::retriable(format!(
+                "statement did not complete within {timeout:?}; discarded the connection and will retry on a fresh one"
+            )))
+        }
     }
 }
 
@@ -416,11 +470,16 @@ pub async fn create_schema_and_table_if_needed(
 
 /// Truncate finalized checkpoint data from the table
 /// This is a best-effort operation - errors are logged but don't fail the pipeline
+///
+/// The DELETE runs inline in the sink's batch loop, so it carries the same
+/// client-side bound as data statements (`execute_bounded`) — an unbounded
+/// await here would wedge the sink exactly like a hung INSERT.
 pub async fn truncate_finalized_checkpoint_data(
     pool: &PgPool,
     schema: &str,
     table: &str,
     finalized_epoch: u64,
+    statement_timeout: Option<Duration>,
 ) -> Result<()> {
     let delete_sql = format!(
         r#"DELETE FROM "{}"."{}" WHERE "_gs_checkpoint_epoch" <= $1"#,
@@ -432,13 +491,9 @@ pub async fn truncate_finalized_checkpoint_data(
         schema, table, finalized_epoch
     );
 
-    match sqlx::query(&delete_sql)
-        .bind(finalized_epoch as i64)
-        .execute(pool)
-        .await
-    {
-        Ok(result) => {
-            let rows_deleted = result.rows_affected();
+    let query = sqlx::query(&delete_sql).bind(finalized_epoch as i64);
+    match execute_bounded(pool, query, statement_timeout).await {
+        Ok(rows_deleted) => {
             info!(
                 "Truncated {} rows from {}.{} for epoch {}",
                 rows_deleted, schema, table, finalized_epoch
@@ -662,5 +717,143 @@ mod tests {
             false,
         );
         assert_eq!(arrow_field_to_postgres_type(&field), "TEXT");
+    }
+
+    use sqlx::postgres::PgPoolOptions;
+    use std::net::SocketAddr;
+    use std::sync::atomic::{AtomicUsize, Ordering};
+    use tokio::io::{AsyncReadExt, AsyncWriteExt};
+    use tokio::net::TcpListener;
+
+    /// Minimal fake PostgreSQL server: completes the startup handshake, then
+    /// never responds to anything else while still reading (and ACKing) the
+    /// client's bytes. This models a peer that died silently mid-connection —
+    /// the client's write succeeds and its read waits forever, which is
+    /// exactly the state that used to hang the sink with no logs.
+    async fn spawn_silent_postgres() -> (SocketAddr, Arc<AtomicUsize>) {
+        let listener = TcpListener::bind("127.0.0.1:0").await.expect("bind");
+        let addr = listener.local_addr().expect("local_addr");
+        let connections = Arc::new(AtomicUsize::new(0));
+        let accepted = connections.clone();
+
+        tokio::spawn(async move {
+            loop {
+                let Ok((mut socket, _)) = listener.accept().await else {
+                    break;
+                };
+                accepted.fetch_add(1, Ordering::SeqCst);
+
+                tokio::spawn(async move {
+                    // Read the client's StartupMessage (length-prefixed, no type byte).
+                    let mut len_buf = [0u8; 4];
+                    if socket.read_exact(&mut len_buf).await.is_err() {
+                        return;
+                    }
+                    let len = (u32::from_be_bytes(len_buf) as usize).saturating_sub(4);
+                    let mut body = vec![0u8; len];
+                    if socket.read_exact(&mut body).await.is_err() {
+                        return;
+                    }
+
+                    let mut reply: Vec<u8> = Vec::new();
+                    // AuthenticationOk
+                    reply.extend([b'R', 0, 0, 0, 8, 0, 0, 0, 0]);
+                    // ParameterStatus messages the client parses during connect
+                    for (key, value) in [
+                        ("server_version", "14.0"),
+                        ("client_encoding", "UTF8"),
+                        ("DateStyle", "ISO, MDY"),
+                    ] {
+                        let payload_len = 4 + key.len() + 1 + value.len() + 1;
+                        reply.push(b'S');
+                        reply.extend((payload_len as u32).to_be_bytes());
+                        reply.extend(key.as_bytes());
+                        reply.push(0);
+                        reply.extend(value.as_bytes());
+                        reply.push(0);
+                    }
+                    // BackendKeyData
+                    reply.extend([b'K', 0, 0, 0, 12]);
+                    reply.extend(1234u32.to_be_bytes());
+                    reply.extend(5678u32.to_be_bytes());
+                    // ReadyForQuery (idle)
+                    reply.extend([b'Z', 0, 0, 0, 5, b'I']);
+                    if socket.write_all(&reply).await.is_err() {
+                        return;
+                    }
+
+                    // Handshake complete. Play dead: keep draining client bytes
+                    // so writes succeed, but never send another byte back.
+                    let mut sink_buf = [0u8; 4096];
+                    while socket
+                        .read(&mut sink_buf)
+                        .await
+                        .map(|n| n > 0)
+                        .unwrap_or(false)
+                    {}
+                });
+            }
+        });
+
+        (addr, connections)
+    }
+
+    /// Reproduces the silent-hang bug: without the client-side bound in
+    /// `execute_bounded`, executing a statement against a peer that stopped
+    /// responding awaits forever (no error, no retry, no log). With the bound
+    /// it must fail retriable within the timeout, discard the dead connection
+    /// (releasing its pool permit), and get a fresh physical connection on
+    /// every subsequent attempt — the third attempt exists to catch a
+    /// permit-leak regression (a plain drop instead of detach strands the
+    /// permit in a background return-to-pool ping and exhausts a 2-connection
+    /// pool by attempt 3).
+    #[tokio::test]
+    async fn test_bounded_execute_times_out_and_discards_dead_connection() {
+        let bound = Duration::from_millis(300);
+        let (addr, connections) = spawn_silent_postgres().await;
+        let url = format!(
+            "postgres://user@{}:{}/db?sslmode=disable",
+            addr.ip(),
+            addr.port()
+        );
+        let pool = PgPoolOptions::new()
+            .max_connections(2)
+            .acquire_timeout(Duration::from_secs(5))
+            .connect_lazy(&url)
+            .expect("lazy pool");
+
+        for attempt in 1..=3u32 {
+            let started = std::time::Instant::now();
+            let result = tokio::time::timeout(
+                Duration::from_secs(10),
+                execute_bounded(&pool, sqlx::query("SELECT 1"), Some(bound)),
+            )
+            .await
+            .unwrap_or_else(|_| panic!("attempt {attempt} must not hang on a silent server"));
+
+            let err = result.expect_err("must time out against a silent server");
+            assert!(err.is_retriable(), "timeout must be retriable: {err:?}");
+            // Distinguish a genuine statement timeout from an acquire failure:
+            // the bound must actually elapse and produce the timeout message.
+            assert!(
+                format!("{err:?}").contains("did not complete within"),
+                "attempt {attempt} must fail via the statement bound, got: {err:?}"
+            );
+            assert!(
+                started.elapsed() >= bound,
+                "attempt {attempt} failed before the bound elapsed: {:?}",
+                started.elapsed()
+            );
+            assert!(
+                started.elapsed() < Duration::from_secs(5),
+                "attempt {attempt} took too long: {:?}",
+                started.elapsed()
+            );
+            assert_eq!(
+                connections.load(Ordering::SeqCst),
+                attempt as usize,
+                "each attempt must open a fresh physical connection (permit released by detach)"
+            );
+        }
     }
 }
